+++ site/src/App.jsx
@@ -9,6 +9,7 @@ import DownloadPage from './pages/DownloadPage'
 import PlanosPage from './pages/PlanosPage'
 import EquipePage from './pages/EquipePage'
 import NovidadesPage from './pages/NovidadesPage'
+import EventosPage from './pages/EventosPage'
 
 function ScrollToTop() {
   const { pathname } = useLocation()
@@ -29,6 +30,7 @@ export default function App() {
           <Route path="/planos" element={<PlanosPage />} />
           <Route path="/equipe" element={<EquipePage />} />
           <Route path="/novidades" element={<NovidadesPage />} />
+          <Route path="/eventos" element={<EventosPage />} />
           <Route path="*" element={<Navigate to="/" replace />} />
         </Routes>
       </main>